Archive 2016 site
move to a subfolder, do some css shuffling and add some trello style guide stylings to help scope for also adding 2017
