+++ src/components/navigation/conference/2016/navigation.scss
@@ -1,0 +1,85 @@
+@import "../../../../colors";
+@import "../../../../frameless";
+
+#navigation {
+    .ul.mod-2016,
+    .li-right-ul.mod-2016 {
+        display: flex;
+        justify-content: space-between;
+        flex-flow: row nowrap;
+        align-items: center;
+        list-style-type: none;
+    }
+
+    .li-left,
+    .li-right,
+    .link {
+        margin-top: 0;
+        margin-right: 10px;
+        color: $type-white;
+    }
+
+    .logo-a {
+        display: flex;
+        height: 100%;
+        align-items: center;
+    }
+
+    .logo-a-image {
+        margin-right: 10px;
+        border-right: 2px solid $active-gray;
+        padding-right: 10px;
+        width: 80px;
+    }
+
+    .logo-a-title {
+        text-decoration: none;
+        white-space: nowrap;
+        color: $type-white;
+        font-size: .85rem;
+        font-weight: bold;
+    }
+
+    @media only screen and (max-width: $mobile - 1) {
+        .li-right-ul.mod-2016 {
+            flex-flow: row nowrap;
+        }
+
+        .link {
+            margin: 0;
+            padding: 0;
+        }
+
+        .link-a {
+            font-size: .75rem;
+        }
+    }
+
+    @media only screen and (max-width: $tablet - 1) {
+        height: 100px;
+
+        .ul.mod-2016 {
+            justify-content: center;
+            flex-flow: row wrap;
+        }
+
+        .li-left,
+        .li-right,
+        .link {
+            align-self: center;
+        }
+
+        .li-right {
+            margin-left: 0;
+            align-self: center;
+        }
+
+        .logo-a {
+            padding-top: 6px;
+        }
+
+        .logo-a-image {
+            padding-left: 15px;
+        }
+    }
+}
