--- src/components/navigation/conference/navigation.scss
@@ -1,94 +1,0 @@
-@import "../../../colors";
-@import "../../../frameless";
-
-#navigation {
-    .inner {
-        ul {
-            display: flex;
-            justify-content: space-between;
-            flex-flow: row nowrap;
-            align-items: center;
-            list-style-type: none;
-
-            li {
-                margin-top: 0;
-                margin-right: 10px;
-                color: $type-white;
-            }
-
-            .logo {
-                > a {
-                    display: flex;
-                    height: 100%;
-                    align-items: center;
-                }
-
-                img {
-                    margin-right: 10px;
-                    border-right: 2px solid $active-gray;
-                    padding-right: 10px;
-                    width: 80px;
-                }
-
-                p {
-                    text-decoration: none;
-                    white-space: nowrap;
-                    color: $type-white;
-                    font-size: .85rem;
-                    font-weight: bold;
-                }
-            }
-        }
-    }
-
-    @media only screen and (max-width: $mobile - 1) {
-        .inner > ul {
-            li {
-                &.right {                                        
-                    ul {
-                        flex-flow: row nowrap;
-
-                        li {
-                            margin: 0;
-                            padding: 0;
-
-                            a {
-                                font-size: .75rem;
-                            }
-                        }
-                    }
-                }
-            }
-        }
-    }
-
-    @media only screen and (max-width: $tablet - 1) {
-        height: 100px;
-
-        .inner {
-            > ul {
-                justify-content: center;
-                flex-flow: row wrap;
-
-                li {
-                    align-self: center;
-
-                    &.right {
-                        margin-left: 0;
-                        align-self: center;
-                    }
-                }
-
-                .logo {
-                    > a {
-                        padding-top: 6px;
-                    }
-                }
-            }
-
-            img {
-                padding-left: 15px;
-            }
-        }
-    }
-}
