+++ src/components/page/conference/2016/page.jsx
@@ -1,15 +1,15 @@
 var React = require('react');
 
-var Navigation = require('../../navigation/conference/navigation.jsx');
-var Footer = require('../../footer/conference/footer.jsx');
+var Navigation = require('../../../navigation/conference/2016/navigation.jsx');
+var Footer = require('../../../footer/conference/footer.jsx');
 
-require('./page.scss');
+require('../page.scss');
 
 var Page = React.createClass({
     type: 'Page',
     render: function () {
         return (
-            <div className="page">
+            <div className="page mod-conference">
                 <div id="navigation">
                     <Navigation />
                 </div>
