+++ src/components/page/conference/page.scss
@@ -29,7 +29,7 @@
     padding: 0;
 }
 
-.index,
+.index.mod-2016,
 .plan,
 .expect,
 .schedule {
@@ -55,52 +55,52 @@
     @media only screen and (max-width: $tablet - 1) {
         margin-top: 50px;
     }
+}
 
-    .title-banner {
-        margin-bottom: 0;
-        background-color: $ui-blue;
-        padding: 2rem 0;
+.title-banner.mod-conference {
+    margin-bottom: 0;
+    background-color: $ui-blue;
+    padding: 2rem 0;
 
-        h1,
-        h4,
-        p {
-            margin: 0 auto;
-            padding: 5px 0;
-            text-align: center;
-            color: $type-white;
-        }
-
-        h1 {
-            font-size: 4.5rem;
-        }
-
-        @media only screen and (max-width: $tablet - 1) {
-            h1 {
-                font-size: 2.5rem;
-            }
-        }
-    }
-
-    .title-icon {
-        position: absolute;
-        width: 100%;
+    h1,
+    h4,
+    p {
+        margin: 0 auto;
+        padding: 5px 0;
         text-align: center;
+        color: $type-white;
+    }
 
+    h1 {
+        font-size: 4.5rem;
+    }
+
+    @media only screen and (max-width: $tablet - 1) {
+        h1 {
+            font-size: 2.5rem;
+        }
+    }
+}
+
+.title-icon {
+    position: absolute;
+    width: 100%;
+    text-align: center;
+
+    img {
+        width: 125px;
+    }
+
+    @media only screen and (max-width: $tablet - 1) {
         img {
-            width: 125px;
-        }
-
-        @media only screen and (max-width: $tablet - 1) {
-            img {
-                transform: translate(0, 5px);
-                width: 85px;
-            }
+            transform: translate(0, 5px);
+            width: 85px;
         }
     }
+}
 
-    section {
-        padding: 64px 0;
-    }
+section {
+    padding: 64px 0;
 }
 
 @media only screen and (max-width: $mobile - 1) {
