+++ src/routes.json
@@ -48,31 +48,39 @@
         "viewportWidth": "device-width"
     },
     {
-        "name": "conference-plan",
-        "pattern": "^/conference/plan/?$",
+        "name": "conference-index-2016",
+        "pattern": "^/conference/2016/?$",
         "routeAlias": "/conference(?!/201[4-5])",
-        "view": "conference/plan/plan",
+        "view": "conference/2016/index/index",
+        "title": "Scratch Conference",
+        "viewportWidth": "device-width"
+    },
+    {
+        "name": "conference-plan-2016",
+        "pattern": "^/conference/2016/plan/?$",
+        "routeAlias": "/conference(?!/201[4-5])",
+        "view": "conference/2016/plan/plan",
         "title": "Plan Your Visit"
     },
     {
-        "name": "conference-expectations",
-        "pattern": "^/conference/expect/?$",
+        "name": "conference-expectations-2016",
+        "pattern": "^/conference/2016/expect/?$",
         "routeAlias": "/conference(?!/201[4-5])",
-        "view": "conference/expect/expect",
+        "view": "conference/2016/expect/expect",
         "title": "What to Expect"
     },
     {
-        "name": "conference-schedule",
-        "pattern": "^/conference/schedule/?$",
+        "name": "conference-schedule-2016",
+        "pattern": "^/conference/2016/schedule/?$",
         "routeAlias": "/conference(?!/201[4-5])",
-        "view": "conference/schedule/schedule",
+        "view": "conference/2016/schedule/schedule",
         "title": "Conference Schedule"
     },
     {
-        "name": "conference-details",
-        "pattern": "^/conference/:id/details/?$",
+        "name": "conference-details-2016",
+        "pattern": "^/conference/2016/:id/details/?$",
         "routeAlias": "/conference(?!/201[4-5])",
-        "view": "conference/details/details",
+        "view": "conference/2016/details/details",
         "title": "Event Details"
     },
     {
