+++ src/views/conference/2016/details/details.scss
@@ -1,4 +1,4 @@
-@import "../../../frameless";
+@import "../../../../frameless";
 
 #view {
     @media only screen and (max-width: $tablet - 1) {
