+++ src/views/conference/2016/expect/expect.scss
@@ -1,5 +1,5 @@
-@import "../../../colors";
-@import "../../../frameless";
+@import "../../../../colors";
+@import "../../../../frameless";
 
 .expect {
     .flex-row {
