+++ src/views/conference/2016/plan/plan.scss
@@ -1,5 +1,5 @@
-@import "../../../colors";
-@import "../../../frameless";
+@import "../../../../colors";
+@import "../../../../frameless";
 
 .plan {
     section {
