+++ src/views/conference/2016/schedule/schedule.scss
@@ -1,5 +1,5 @@
-@import "../../../colors";
-@import "../../../frameless";
+@import "../../../../colors";
+@import "../../../../frameless";
 
 .schedule {
     .title-banner {
